add confirmation before deleteing the profile

+++ lib/screens/profil/profil_details.dart
@@ -38,24 +38,51 @@ class _ProfilDetailsState extends State<ProfilDetails> {
           IconButton(
             icon: Icon(Icons.delete),
             onPressed: () async {
-              final bool success = await ProfilService().deleteProfil(widget.profil.id);
-              if (success) {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text('Profil deleted successfully'),
-                    backgroundColor: Colors.green,
-                  ),
-                );
-                Navigator.pop(context);
-                print('onProfileUpdated callback is being called'); // Add this print statement
-                widget.onProfileUpdated(); // Call the onProfileUpdated method
-              } else {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text('Failed to delete profil'),
-                    backgroundColor: Colors.red,
-                  ),
-                );
+              final bool? confirmDelete = await showDialog<bool>(
+                context: context,
+                builder: (BuildContext context) {
+                  return AlertDialog(
+                    title: const Text('Confirm Delete'),
+                    content: const Text(
+                        'Are you sure you want to delete this profile?'),
+                    actions: <Widget>[
+                      TextButton(
+                        child: const Text('Cancel'),
+                        onPressed: () {
+                          Navigator.of(context).pop(false);
+                        },
+                      ),
+                      TextButton(
+                        child: const Text('Delete'),
+                        onPressed: () {
+                          Navigator.of(context).pop(true);
+                        },
+                      ),
+                    ],
+                  );
+                },
+              );
+
+              if (confirmDelete == true) {
+                final bool success =
+                    await ProfilService().deleteProfil(widget.profil.id);
+                if (success) {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(
+                      content: Text('Profil deleted successfully'),
+                      backgroundColor: Colors.green,
+                    ),
+                  );
+                  Navigator.pop(context);
+                  widget.onProfileUpdated();
+                } else {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(
+                      content: Text('Failed to delete profil'),
+                      backgroundColor: Colors.red,
+                    ),
+                  );
+                }
               }
             },
           ),
